requirements.txt
langchain
langchain-google-genai
langgraph
langchain-community
requests
feedparser
reportlab
python-dotenv
pydantic